requirements.txt
biopython>=1.74
deprecation>=2.1.0
domdf_python_tools>=0.12.0
enum_tools>=0.6.1
matplotlib>=3.0.2
netCDF4>=1.5.0
numpy>=1.16.2
openpyxl>=2.6.2
pandas>=1.0.0
pymzml>=2.2.1
scipy>=1.2.1
sdjson>=0.3.0
